Async: support multiple parallel requests in a single Kafka::Connection
We will very likely be reading more than one partition and if that
happens there is a high chance that we are connected already to the
right server and that this connection is not currently in use.
With this change I make sure the Kafka::Connection can handle multiple
simultaneous connections to the same server. I also make sure the
connections that are currently in use are correctly marked, so we will
not try to throw multiple requests at them.